+++ sccfm_cli/commands/inventory/devices/cdfmc_managed_ftd/cli/__init__.py
@@ -1,0 +1,5 @@
+from sccfm_cli.commands.inventory.devices.cdfmc_managed_ftd.cli.command import (
+    FtdCliCommand,
+)
+
+__all__ = ["FtdCliCommand"]