+++ sccfm_cli/commands/inventory/devices/cdfmc_managed_ftd/cli/execute/__init__.py
@@ -1,0 +1,5 @@
+from sccfm_cli.commands.inventory.devices.cdfmc_managed_ftd.cli.execute.command import (
+    FtdExecuteCliCommand,
+)
+
+__all__ = ["FtdExecuteCliCommand"]