+++ sccfm_core/services/__init__.py
@@ -10,6 +10,7 @@ from sccfm_core.services.inventory import (
     AsaUpgradeService,
     AsaUpgradeVersionService,
     AsaUserPasswordService,
+    FtdCommandLineService,
     InventoryService,
     ShunEntrySpec,
 )
@@ -41,6 +42,7 @@ __all__ = [
     "AsaUpgradeService",
     "AsaUpgradeVersionService",
     "AsaUserPasswordService",
+    "FtdCommandLineService",
     "HealthService",
     "HealthStatus",
     "InventoryService",